# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "tarefas for me",
        "description": "APIs são baseadas na arquitetura REST e usam os métodos básicos de solicitação HTTP. Para melhor compreensão, os recursos são listados juntamente com seus parâmetros obrigatórios, se houver, e seus exemplos de solicitações e respostas logo abaixo da descrição. As APIs aceitam solicitações de corpo codificado em JSON e retornam dados da mesma forma.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/labels": {
            "get": {
                "tags": [
                    "Label"
                ],
                "summary": "Labels",
                "operationId": "2ac7cbc4186e08de3ec96a7898844f10",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "example": 200
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "content": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "userId": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-04-07T23:15:09.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-04-07T23:15:09.000000Z"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Label Name"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Label description"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "#FFFFFF"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Label"
                ],
                "summary": "Create new label",
                "operationId": "02bc8c47faba31d6682c4bafd253f42c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "color": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Urgente",
                                    "description": "Description",
                                    "color": "#FFFFFF"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "example": 200
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "content": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "userId": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-04-07T23:15:09.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-04-07T23:15:09.000000Z"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Label Name"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Label description"
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "#FFFFFF"
                                                }
                                            },
                                            "type": "object",
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/labels/{id}": {
            "put": {
                "tags": [
                    "Label"
                ],
                "summary": "Update label",
                "operationId": "e7b8ab515f4ba489ed5d3bfe4b9b354b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of label",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "color": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Urgente",
                                    "description": "Description",
                                    "color": "#FFFFFF"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "example": 200
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "content": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "userId": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-04-07T23:15:09.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-04-07T23:15:09.000000Z"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Label Name"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Label description"
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "#FFFFFF"
                                                }
                                            },
                                            "type": "object",
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Label"
                ],
                "summary": "Delete label",
                "operationId": "2e8f90cafd1d857e5da061c85c627b00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of label",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "properties": {
                                                "code": {
                                                    "type": "number",
                                                    "example": 200
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "success"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "oauth2",
                "name": "Authorization"
            }
        }
    }
}